requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.23.4
PyPDF2==3.0.1
python-docx==1.1.0
groq==0.4.2
apscheduler==3.10.4
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
